sesson 28

+++ Modules/Mlk/Category/Http/Controllers/CategoryController.php
@@ -4,6 +4,7 @@ namespace Mlk\Category\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use Mlk\Category\Http\Requests\CategoryRequest;
+use Mlk\Category\Models\Category;
 use Mlk\Category\Repositories\CategoryRepo;
 use Mlk\Category\Services\CategoryService;
 
@@ -20,24 +21,31 @@ class CategoryController extends Controller
 
     public function index()
     {
+        $this->authorize('index', Category::class);
         $categories = $this->repo->index()->paginate(10);
+
         return view('Category::index', compact('categories'));
     }
 
     public function create()
     {
+        $this->authorize('index', Category::class);
         $categories = $this->repo->index()->get();
+
         return view('Category::create', compact('categories'));
     }
 
     public function store(CategoryRequest $request)
     {
+        $this->authorize('index', Category::class);
         $this->service->store($request);
+
         return to_route('categories.index')->with(['success_message' => 'دسته بندی با موفقیت ساخته شد']);
     }
 
     public function edit($id)
     {
+        $this->authorize('index', Category::class);
         $category = $this->repo->findById($id);
         $categories = $this->repo->index()->where('id', '!=', $category->id)->get();
 
@@ -46,18 +54,23 @@ class CategoryController extends Controller
 
     public function update(CategoryRequest $request, $id)
     {
+        $this->authorize('index', Category::class);
         $this->service->update($request, $id);
+
         return to_route('categories.index')->with(['success_message' => 'دسته بندی با موفقیت ابدیت شد']);
     }
 
     public function destroy($id)
     {
+        $this->authorize('index', Category::class);
         $this->repo->delete($id);
+
         return to_route('categories.index')->with(['success_message' => 'دسته بندی با موفقیت حذف شد']);
     }
 
     public function changeStatus($id)
     {
+        $this->authorize('index', Category::class);
         $category = $this->repo->findById($id);
         $this->repo->changeStatus($category);
