+++ Modules/Mlk/User/Http/Controllers/UserController.php
@@ -7,6 +7,7 @@ use Mlk\Role\Repositories\RoleRepo;
 use Mlk\User\Http\Requests\AddRoleRequest;
 use Mlk\User\Http\Requests\UserRequest;
 use Mlk\User\Http\Requests\UserUpdateRequest;
+use Mlk\User\Models\User;
 use Mlk\User\Repositories\UserRepo;
 use Mlk\User\Services\UserService;
 
@@ -30,48 +31,62 @@ class UserController extends Controller
 
     public function index()
     {
+        $this->authorize('index', User::class);
         $users = $this->repo->index();
+
         return view('User::index', compact('users'));
     }
 
     public function create()
     {
+        $this->authorize('index', User::class);
         return view('User::create');
     }
 
     public function store(UserRequest $request)
     {
+        $this->authorize('index', User::class);
         $this->service->store($request);
+
         return to_route('users.index');
     }
 
     public function edit($id)
     {
+        $this->authorize('index', User::class);
         $user = $this->repo->findById($id);
+
         return view('User::edit', compact('user'));
     }
 
     public function update(UserUpdateRequest $request, $id)
     {
+        $this->authorize('index', User::class);
         $this->service->update($request, $id);
+
         return to_route('users.index');
     }
 
     public function destroy($id)
     {
+        $this->authorize('index', User::class);
         $this->repo->delete($id);
+
         return to_route('users.index')->with(['success_delete' => 'کاربر با موفقیت حذف شد']);
     }
 
     // Role
     public function addRole($user_id, RoleRepo $roleRepo)
     {
+        $this->authorize('index', User::class);
         $roles = $roleRepo->index()->get();
+
         return view('User::add-roles', compact(['user_id', 'roles']));
     }
 
     public function addRoleStore(AddRoleRequest $request, $userId)
     {
+        $this->authorize('index', User::class);
         $user = $this->repo->findById($userId);
         $this->service->addRole($request->role, $user);
 
@@ -81,6 +96,7 @@ class UserController extends Controller
 
     public function removeRole($userId, $roleId, RoleRepo $roleRepo)
     {
+        $this->authorize('index', User::class);
         $user = $this->repo->findById($userId);
         $role = $roleRepo->findById($roleId);
         $this->service->deleteRole($user, $role->name);
